pyproject.toml
[project]
name = "decipher-news-agent"
version = "0.1.0"
description = "DNA - Personal AI agent that generates a daily newspaper"
requires-python = ">=3.13"
dependencies = [
    "google-auth==2.38.0",
    "google-auth-oauthlib==1.2.1",
    "requests==2.32.3",
    "pyyaml==6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]
